SPAR-2838 Update WDIO to 9.x and jinja2 to 3.1.5 (#1882)
* Install taurus_wdio_plugin and change cwd inside tools for wdio

* Install taurus_wdio_plugin and change cwd inside tools for wdio 2 - npm copy package not link

* Remove dev files

* Remove common js variant of rewritten files (to mjs)

* Collect all results in WDIO custom reporter not just last run

* Change WDIO tools dir as cjs and mjs in same node_modules are causing problems

Also for case we will need to revert back to old wdio, it will be safer to have old tools dir untouched

* Testability - cwd from method

* Do not break others because of wdio change from cjs to mjs

---------
