+++ bzt/resources/wdio-taurus-plugin/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@taurus/wdio-taurus-plugin",
+  "version": "1.0.0",
+  "description": "",
+  "main": "wdio-taurus-plugin.mjs",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "type": "module",
+  "dependencies": {
+    "@wdio/cli": "^9.2.1",
+    "@wdio/json-reporter": "^9.1.3",
+    "@wdio/local-runner": "^9.2.1",
+    "@wdio/mocha-framework": "^9.1.3"
+  }
+}
