+++ tests/unit/modules/_selenium/test_javascript.py
@@ -166,6 +166,7 @@ class TestWebdriverIOExecutor(SeleniumTestCase):
     def full_run(self, config):
         self.prepare(config)
         self.obj.runner.get_launch_cmdline = lambda *args: [TestWebdriverIOExecutor.RUNNER_STUB] + list(args)
+        self.obj.runner.get_launch_cwd = lambda *args: None
         self.obj.startup()
         while not self.obj.check():
             time.sleep(self.obj.engine.check_interval)
